requirements.txt
appdirs==1.4.4
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
cssselect==1.2.0
dateparser==1.1.4
debugpy==1.6.4
decorator==5.1.1
demjson3==3.0.6
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
facebook-scraper==0.2.59
fake-useragent==1.1.1
Flask==2.2.2
geckodriver-autoinstaller==0.1.0
h11==0.14.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.0
outcome==1.2.0
packaging==22.0
pandas==1.5.2
parse==1.19.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyee==8.2.2
Pygments==2.13.0
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-html==0.10.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
w3lib==2.1.1
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
